Lima Windows support

+++ pkg/store/instance.go
@@ -215,16 +215,18 @@ func ReadPIDFile(path string) (int, error) {
 	if runtime.GOOS == "windows" {
 		return pid, nil
 	}
-	err = proc.Signal(syscall.Signal(0))
-	if err != nil {
-		if errors.Is(err, os.ErrProcessDone) {
-			_ = os.Remove(path)
-			return 0, nil
-		}
-		// We may not have permission to send the signal (e.g. to network daemon running as root).
-		// But if we get a permissions error, it means the process is still running.
-		if !errors.Is(err, os.ErrPermission) {
-			return 0, err
+	if runtime.GOOS != "windows" {
+		err = proc.Signal(syscall.Signal(0))
+		if err != nil {
+			if errors.Is(err, os.ErrProcessDone) {
+				_ = os.Remove(path)
+				return 0, nil
+			}
+			// We may not have permission to send the signal (e.g. to network daemon running as root).
+			// But if we get a permissions error, it means the process is still running.
+			if !errors.Is(err, os.ErrPermission) {
+				return 0, err
+			}
 		}
 	}
 	return pid, nil
